requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
opencv-python>=4.8.0
pillow>=10.0.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0

# Machine Learning Metrics
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0
pandas>=2.0.0

# Optional: For better performance
# tensorboard>=2.14.0  # Uncomment for training visualization



